part_5_production/chapter_18_testing/requirements.txt
# requirements.txt
pytest==9.0.2
pytest-asyncio==1.3.0
langgraph==1.0.4
langchain-core==1.2.0
langchain-openai==1.1.1
python-dotenv==1.2.1
httpx==0.28.1
numpy==2.3.5
scipy==1.16.3